Create separate css files for each component
